Update svg2ttf and freeze timestamp
The TTF spec requires that a timestamp be embedded in each file.
svg2ttf follows this and generates a new one every time it compiles
a new file. This is problematic for caching based on file contents.
Even if no source files change, the bytes in the created ttf file
will (because of the changed timestamp). A `ts` option was recently
added, enabling us to specify a static timestamp to prevent this.